+++ catbuffer/parser/CHANGELOG.md
@@ -3,6 +3,20 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [3.0.0] - 17-Dec-2021
+
+### Added
+- struct attributes: {size, initializes, discriminator}
+- struct field attributes: {is_byte_constrained, alignment, sort_key}
+- struct "abstract" modifier
+- AST object model
+- CATS DSL documentation
+
+### Changed
+- rewrite entire parser using Lark grammar
+- array(...) syntax
+- deprecated YAML output format
+
 ## [2.0.2] - 15-Sep-2021
 
 ### Fixed
@@ -41,6 +55,7 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 ### Added
  - initial code release
 
+[3.0.0]: https://github.com/symbol/catbuffer-parser/compare/v2.0.2...v3.0.0
 [2.0.2]: https://github.com/symbol/catbuffer-parser/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/symbol/catbuffer-parser/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/symbol/catbuffer-parser/compare/v1.0.0...v2.0.0